[feat/#64] 타운옵션버튼 구현
